requirements.txt
Flask~=2.2.2
flask_sqlalchemy
flask_login
flask_wtf
Flask-Talisman
pyotp==2.8.0
WTForms~=3.0.1
email_validator
python-dotenv~=0.21.0
SQLAlchemy~=1.4.41
bcrypt~=4.0.1
cryptography~=38.0.4

MarkupSafe~=2.1.1